# Use a set to track constrained pairs in OpenMM export