--- a/openff/interchange/tests/data/tip4p-ew.offxml
+++ b/openff/interchange/tests/data/tip4p-ew.offxml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding='ASCII'?>
+<SMIRNOFF version="0.3" aromaticity_model="OEAroModel_MDL">
+  <vdW version="0.3" potential="Lennard-Jones-12-6" combining_rules="Lorentz-Berthelot" scale12="0.0" scale13="0.0" scale14="0.5" scale15="1" switch_width="1.0 * angstroms" cutoff="9.0 * angstroms" method="cutoff">
+    <Atom smirks="[#1]-[#8X2H2+0:1]-[#1]" id="n1" sigma="0.316435 * nanometers" epsilon="0.680946 * kilojoules_per_mole" />
+    <Atom smirks="[#1:1]-[#8X2H2+0]-[#1]" id="n2" sigma="1 * nanometers" epsilon="0 * kilojoules_per_mole" />
+  </vdW>
+  <LibraryCharges version="0.3">
+    <LibraryCharge name="TIP4P" smirks="[#1:1]-[#8X2H2+0:2]-[#1:3]" charge1="0.0 * elementary_charge" charge2="0.0 * elementary_charge" charge3="0.0 * elementary_charge"/>
+  </LibraryCharges>
+  <VirtualSites version="0.3" exclusion_policy="parents">
+    <VirtualSite
+      type="DivalentLonePair"
+      name="EP"
+      smirks="[#1:2]-[#8X2H2+0:1]-[#1:3]"
+      distance="-0.125 * angstrom"
+      charge_increment1="0.52422 * elementary_charge"
+      charge_increment2="0.0 * elementary_charge"
+      charge_increment3="0.52422 * elementary_charge"
+      sigma="1.0 * angstrom"
+      epsilon="0.0 * kilocalories_per_mole"
+      outOfPlaneAngle="0 * degree"
+      match="once">
+    </VirtualSite>
+  </VirtualSites>
+  <Constraints version="0.3">
+    <Constraint smirks="[#1:1]-[#8X2H2+0:2]-[#1]" id="c1" distance="0.9572 * angstrom"/>
+    <Constraint smirks="[#1:1]-[#8X2H2+0]-[#1:2]" id="c2" distance="1.5139006545247014 * angstrom"/>
+  </Constraints>
+</SMIRNOFF>