### Description
While running `Interchange.to_openmm` on a protein-ligand complex, a sizable amount of runtime is spent inside of `_is_constrained`, which iterated through a list of tuples (of ints) to see if any of the elements matched a particular tuple. This list served no other purpose (ordering didn't matter, it never needed to be indexed, etc.) so it's a canonical case in a a `set` is a better choice. Using a recent version of the [toolkit showcase](https://github.com/openforcefield/openff-toolkit/blob/0.11.0/examples/toolkit_showcase/toolkit_showcase.ipynb) as extremely rough benchmarking, this decreases the runtime by roughly 60% by reducing the list lookup to almost nothing.


```
%%snakeviz
omm_system = interchange.to_openmm()

# BEFORE:
# total: 24.8 seconds,
#    _process_angle_forces 11.4s
#    _is_constrained (total) 14.8s
#
# AFTER:
# total: 10.1
#    _process_angle_forces 0.5s
#    _is_constrained (total) 0.1s
```

### Checklist
- [ ] Add tests
- [ ] Lint
- [ ] Update docstrings